inherit run method from base sim
