+++ maria/noise.py
@@ -11,7 +11,7 @@ class NoiseSimulation(base.BaseSimulation):
 
         self.white_noise_level = kwargs.get("white_noise", 1)
 
-    def run(self):
+    def _run(self):
 
         self.temperature = self.white_noise_level * np.random.standard_normal(size=(self.pointing.nt, self.pointing.nt))
